added scheduler

+++ src/ca/team2706/vision/trackerboxreloaded/Main.java
@@ -44,9 +44,11 @@ public class Main {
 	public static boolean showMiddle = false;
 	public static boolean useCamera = true;
 	public static Mat frame;
-	public static void setFrame(Mat f){
+
+	public static void setFrame(Mat f) {
 		frame = f;
 	}
+
 	// Camera Type (set in visionParams.properties)
 	// Set to 1 for USB camera, set to 0 for webcam, I think 0 is USB if
 	// there is no webcam :/
@@ -395,41 +397,47 @@ public class Main {
 		return mat;
 	}
 
-/**
- * 
- * @param The
- *            image to dump to a file
- * @param image
- *            the image to be dumped
- * @param suffix
- *            the suffix to put on the file name
- * @throws IOException
- */
+	/**
+	 * 
+	 * @param The
+	 *            image to dump to a file
+	 * @param image
+	 *            the image to be dumped
+	 * @param suffix
+	 *            the suffix to put on the file name
+	 * @throws IOException
+	 */
 
-	public static void imgDump(BufferedImage image, String suffix) throws IOException {
+	public static void imgDump(BufferedImage image, String suffix, int timestamp) throws IOException {
 		// prepend the file name with the tamestamp integer, left-padded with
 		// zeros so it sorts properly
-		File output = new File(outputPath + String.format("%05d",timestamp) + "_" + suffix + ".png");
+		File output = new File(outputPath + String.format("%05d", timestamp) + "_" + suffix + ".png");
+		boolean recurse = false;
 		try {
-			if(output.exists()){
+			if (output.exists()) {
+				recurse = true;
 				timestamp++;
-				imgDump(image, suffix);
+				imgDump(image, suffix, timestamp);
 			}
-			ImageIO.write(image, "PNG", output);
+			if (!recurse) {
+				ImageIO.write(image, "PNG", output);
+			}
 		} catch (IOException e) {
 			throw new IOException(e.getMessage());
 		}
-		timestampfile.delete();
-		timestampfile.createNewFile();
-		PrintWriter out = new PrintWriter(timestampfile);
-		out.println(timestamp);
-		out.close();
+		if (!recurse) {
+			timestampfile.delete();
+			timestampfile.createNewFile();
+			PrintWriter out = new PrintWriter(timestampfile);
+			out.println(timestamp);
+			out.close();
+		}
 	}
 
 	/**
 	 * The main method! Very important Do not delete! :] :]
 	 *
-
+	 * 
 	 * @param The
 	 *            command line arguments
 	 */
@@ -445,10 +453,10 @@ public class Main {
 		loadVisionParams();
 
 		try {
-// Copys the vision parameters to a usb flash drive
+			// Copys the vision parameters to a usb flash drive
 			Files.copy(Paths.get("visionParams.properties"),
 					Paths.get(outputPath + "/visionParams-" + timestamp + ".properties"),
-		StandardCopyOption.REPLACE_EXISTING);	
+					StandardCopyOption.REPLACE_EXISTING);
 		} catch (IOException e2) {
 			e2.printStackTrace();
 		}
@@ -525,7 +533,7 @@ public class Main {
 				// non-fatal error, let the program continue
 			}
 		}
-
+		ImageDumpScheduler.start();
 		// Main video processing loop
 		while (true) {
 			if (useCamera) {
@@ -592,36 +600,29 @@ public class Main {
 					continue;
 				}
 			}
-			if(useCamera){
-			// log images to file once every seconds_between_img_dumps
-			double elapsedTime = ( (double) System.currentTimeMillis() / 1000) - current_time_seconds;
-			// If the elapsed time is more that the seconds between image dumps
-			// then dump images asynchronously
-			if (elapsedTime >= seconds_between_img_dumps) {
-				// Sets the current number of seconds
-				current_time_seconds = (((double) System.currentTimeMillis()) / 1000);
-				// Clones the frame
-				Mat finalFrame = frame.clone();
-				// Starts a new thread to dump images
-				new Thread(new Runnable() {
-					public void run() {
-						try {
-							// Dumps the raw image
-							imgDump(matToBufferedImage(finalFrame), "raw");
-							// Dumps the binMask image
-							imgDump(matToBufferedImage(visionData.binMask), "binMask");
-							// Draw the target to the output image
-							Pipeline.drawPreferredTarget(finalFrame, visionData);
-							// Dumps the output image
-							imgDump(matToBufferedImage(finalFrame), "output");
-							timestamp++;
-						} catch (IOException e) {
-							e.printStackTrace();
-							return;
-						}
+			if (useCamera) {
+				// log images to file once every seconds_between_img_dumps
+				double elapsedTime = ((double) System.currentTimeMillis() / 1000) - current_time_seconds;
+				// If the elapsed time is more that the seconds between image
+				// dumps
+				// then dump images asynchronously
+				if (elapsedTime >= seconds_between_img_dumps) {
+					// Sets the current number of seconds
+					current_time_seconds = (((double) System.currentTimeMillis()) / 1000);
+					// Clones the frame
+					Mat finalFrame = frame.clone();
+					try {
+						Mat draw = finalFrame.clone();
+						Pipeline.drawPreferredTarget(draw, visionData);
+						Bundle b = new Bundle(matToBufferedImage(finalFrame), matToBufferedImage(visionData.binMask),
+								matToBufferedImage(draw), timestamp);
+						ImageDumpScheduler.schedule(b);
+						timestamp++;
+					} catch (IOException e) {
+						e.printStackTrace();
+						return;
 					}
-				}).start();
-			}
+				}
 			}
 			// Display the frame rate onto the console
 			double pipelineTime = (((double) (pipelineEnd - pipelineStart)) / Pipeline.NANOSECONDS_PER_SECOND) * 1000;
@@ -659,10 +660,9 @@ public class Main {
 		return visionData;
 	}
 
-	public static Mat getFrame(){
-    	Mat frame = new Mat();
-    	camera.read(frame);
-    	return frame;
-    }
+	public static Mat getFrame() {
+		Mat frame = new Mat();
+		camera.read(frame);
+		return frame;
+	}
 }
-
